pyproject.toml
[tool.poetry]
name = "product-checker"
version = "0.1.0"
description = "Give product info from its photo"
authors = ["Vadim <sharvadim07@yandex.ru>"]
readme = "README.md"
packages = [{include = "product_checker"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "^1.24.2"
scikit-image = "^0.20.0"
opencv-python = "^4.7.0.72"
pytesseract = "^0.3.10"
deskew = "^1.3.3"
scikit-learn = "^1.2.1"
python-dateutil = "^2.8.2"
tqdm = "^4.65.0"
urllib3 = "^1.26.14"
minio = "^7.1.13"

[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
